+++ proxy.hs
@@ -61,8 +61,6 @@ doProcessPage ref uri params = do
 
 			log_page_result ref status_before_func log_time state_before uri params effuri pagetext status_after_func state_after
 
-			forkIO_ "proxy:checkserverstate" $ checkServerState ref
-
 			return (y, pagetext, effuri, hdrs, code)
 		putMVar mv =<< case x of
 			Right (Right msg, _, effuri, hdrs, code) -> do
@@ -143,7 +141,7 @@ make_ref baseref = do
 			apixf <- load_api_status_to_mv_mkapixf ref
 			load_api_status_to_mv ref mv apixf
 		force_latest_status_parse ref
-		void $ loadState ref
+		ensureLoadedState ref
 		return True) `catch` (\e -> do
 			putWarningStrLn $ "loadstate exception: " ++ show (e :: SomeException)
 			return False)
@@ -192,9 +190,7 @@ kolProxyHandler uri params baseref = do
 				putInfoStrLn $ "login.php -> getting server state"
 				ai <- getApiInfo newref
 				putInfoStrLn $ "Logging in as " ++ (charName ai) ++ " (ascension " ++ (show $ ascension ai) ++ ")"
-				what <- loadSettingsFromServer newref
-				putInfoStrLn $ "settings loaded: " ++ (fromMaybe "nothing" what)
-				writeIORef (have_logged_in_ref_ $ globalstuff_ $ newref) True
+				loadSettingsFromServer newref
 
 				forkIO_ "proxy:compresslogs" $ compressLogs (charName ai) (ascension ai)
 
@@ -255,11 +251,9 @@ kolProxyHandler uri params baseref = do
 	retresp <- log_time_interval origref ("run handler for: " ++ (show uri)) $ case response of
 		Just r -> r
 		Nothing -> do
-			canread_before <- canReadState origref
-
-			-- TODO: Move to specific handler? Remove entirely?
-			when (uriPath uri == "/logout.php" && canread_before) $ storeSettingsOnServer origref "logging out"
-
+			when (uriPath uri == "/logout.php") $ do
+				canread_before <- canReadState origref
+				when canread_before $ storeSettings origref
 			let reqtype = if isJust params then "POST" else "GET"
 			response <- log_time_interval origref ("browser request: " ++ (show uri)) $ runBrowserRequestScript origref uri allparams reqtype
 			case response of
@@ -307,7 +301,7 @@ runbot filename = do
 	let login_useragent = kolproxy_version_string ++ " (" ++ platform_name ++ ")" ++ " BotScript/0.1 (" ++ filename ++ ")"
 	let login_host = fromJust $ parseURI $ "http://www.kingdomofloathing.com/"
 
-	sc <- make_sessionconn globalref "http://www.kingdomofloathing.com/" (error "dblogstuff") (error "statestuff")
+	sc <- make_sessionconn globalref "http://www.kingdomofloathing.com/" (error "dblogstuff")
 
 	Just username <- getEnvironmentSetting "KOLPROXY_BOTSCRIPT_USERNAME"
 	Just passwordmd5hash <- getEnvironmentSetting "KOLPROXY_BOTSCRIPT_PASSWORDMD5HASH"
